pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Prometheus"
version = "0.0.1"
dependencies = [
  "langchain==0.3.3",
  "tree-sitter==0.21.3",
  "tree-sitter-languages==1.10.2",
  "neo4j==5.20.0",
  "fastapi[standard]>=0.115.2",
  "langchain-anthropic==0.3.0",
  "langchain-openai==0.2.8",
  "langchain-google-genai==2.0.4",
  "langchain_community==0.3.2",
  "igittigitt>=2.1.5",
  "litellm>=1.52.9",
  "GitPython>=3.1.43",
  "langgraph==0.2.41",
  "langgraph-checkpoint-postgres>=2.0.2",
  "psycopg[binary]>=3.2.3",
  "dynaconf>=3.2.6",
  "docker>=7.1.0",
  "unidiff>=0.7.5",
  "argon2-cffi>=23.1.0",
  "sqlmodel==0.0.24",
  "psycopg2-binary",
  "pyjwt==2.6.0",
]
requires-python = ">= 3.11"

[project.optional-dependencies]
test = [
  "ruff>=0.9.1",
  "anyio",
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
  "pytest-asyncio==1.1.0",
  "pytest-tornasync",
  "pytest-trio",
  "pytest-twisted",
  "twisted",
  "testcontainers==4.8.2",
]

[tool.ruff]
indent-width = 4
line-length = 100

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.pydocstyle]
convention = "google"
